requirements.txt
allure-python-commons
appdirs
astroid
attrs
beautifulsoup4
certifi
charset-normalizer
dill
et-xmlfile
exceptiongroup
idna
iniconfig
isort
jproperties
lazy-object-proxy
lxml
mccabe
openpyxl
packaging
platformdirs
pluggy
pylint
pyparsing
pytest
python-dotenv
PyYAML
requests
selenium==3.14.0
six
soupsieve
tomli
tomlkit
tqdm
typing_extensions
urllib3
webdriver-manager
wrapt